--- src/components/DropZoneDemo/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "DropZoneDemo",
-  "main": "./DropZoneDemo.js"
-}
