+++ src/container/App/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "App",
+  "main": "./App.js"
+}
