+++ src/container/Html/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "Html",
+  "main": "./Html.js"
+}
