+++ src/container/Main/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "Main",
+  "main": "./Main.js"
+}
