+++ src/server.js
@@ -20,8 +20,8 @@ import ReactDOM from 'react-dom/server';
 import { Provider } from 'react-redux';
 import UniversalRouter from 'universal-router';
 import PrettyError from 'pretty-error';
-import App from './components/App';
-import Html from './components/Html';
+import App from './container/App';
+import Html from './container/Html';
 import { ErrorPageWithoutStyle } from './routes/error/ErrorPage';
 import errorPageStyle from './routes/error/ErrorPage.css';
 import passport from './core/passport';
